chore: prune 103 redundant/trivial tests from regular suite
Removed 103 tests (5,505 → 5,402) and deleted 3 test files:
- tests/services/test_provider_clients.py (51 tests — mega file duplicating individual provider files)
- tests/services/test_modelz_client.py (2 tests — 100% hasattr checks)
- tests/services/test_near_client.py (2 tests — 100% hasattr checks)

Pruned from 16 files:
- 12 duplicate tests from test_security.py (covered by CM-01)
- 10 hasattr tests from test_tools.py
- 5 trivial tests from test_statsig_service.py
- 4 trivial tests from test_security_validators.py
- 3 duplicates from test_health.py (covered by CM-11)
- 2 duplicates from test_pricing.py (covered by CM-06)
- 2 duplicates from test_stream_normalizer.py (covered by CM-10)
- 2 trivial tests from test_routing_refactor.py
- 1 duplicate from test_rate_limiting.py (covered by CM-02)
- Provider hasattr tests from 6 files (canopywave, sybil, morpheus, novita, helicone, xai)
- 10 unique Cerebras tests moved from mega file to test_cerebras_client.py

CM tests: 194 passed, 0 xfailed (unchanged)
