+++ tests/services/test_helicone_client.py
@@ -7,27 +7,6 @@ from unittest.mock import Mock, patch
 import pytest
 
 
-class TestHeliconeClient:
-    """Test Helicone Client service functionality"""
-
-    def test_module_imports(self):
-        """Test that module imports successfully"""
-        import src.services.helicone_client
-
-        assert src.services.helicone_client is not None
-
-    def test_module_has_expected_attributes(self):
-        """Test module exports"""
-        from src.services import helicone_client
-
-        assert hasattr(helicone_client, "get_helicone_client")
-        assert hasattr(helicone_client, "make_helicone_request_openai")
-        assert hasattr(helicone_client, "make_helicone_request_openai_stream")
-        assert hasattr(helicone_client, "process_helicone_response")
-        assert hasattr(helicone_client, "fetch_model_pricing_from_helicone")
-        assert hasattr(helicone_client, "get_provider_pricing_for_helicone_model")
-
-
 class TestGetHeliconeClient:
     """Test get_helicone_client function"""
