--- tests/services/test_modelz_client.py
@@ -1,23 +1,0 @@
-"""
-Comprehensive tests for Modelz Client service
-"""
-
-from unittest.mock import AsyncMock, MagicMock, Mock, patch
-
-import pytest
-
-
-class TestModelzClient:
-    """Test Modelz Client service functionality"""
-
-    def test_module_imports(self):
-        """Test that module imports successfully"""
-        import src.services.modelz_client
-
-        assert src.services.modelz_client is not None
-
-    def test_module_has_expected_attributes(self):
-        """Test module exports"""
-        from src.services import modelz_client
-
-        assert hasattr(modelz_client, "__name__")
