--- tests/services/test_near_client.py
@@ -1,23 +1,0 @@
-"""
-Comprehensive tests for Near Client service
-"""
-
-from unittest.mock import AsyncMock, MagicMock, Mock, patch
-
-import pytest
-
-
-class TestNearClient:
-    """Test Near Client service functionality"""
-
-    def test_module_imports(self):
-        """Test that module imports successfully"""
-        import src.services.near_client
-
-        assert src.services.near_client is not None
-
-    def test_module_has_expected_attributes(self):
-        """Test module exports"""
-        from src.services import near_client
-
-        assert hasattr(near_client, "__name__")
