+++ tests/services/test_xai_client.py
@@ -2,25 +2,7 @@
 Comprehensive tests for Xai Client service
 """
 
-from unittest.mock import AsyncMock, MagicMock, Mock, patch
-
-import pytest
-
-
-class TestXaiClient:
-    """Test Xai Client service functionality"""
-
-    def test_module_imports(self):
-        """Test that module imports successfully"""
-        import src.services.xai_client
-
-        assert src.services.xai_client is not None
-
-    def test_module_has_expected_attributes(self):
-        """Test module exports"""
-        from src.services import xai_client
-
-        assert hasattr(xai_client, "__name__")
+from unittest.mock import MagicMock, patch
 
 
 class TestXaiReasoningDetection:
